move mainWarning to reusable component for all rails
